requirements.txt
python-docx
requests
beautifulsoup4
openpyxl
PyPDF2
